dbank-BE/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.5
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==1.6.4
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
